Docs, Tests, Refactor
