+++ .circleci/config.yml
@@ -39,39 +39,32 @@ jobs:
   doctest:
     <<: *common
     docker:
-      - image: circleci/python:3.5
+      - image: circleci/python:3.6
         environment:
           TOXENV: doctest
   lint:
     <<: *common
     docker:
-      - image: circleci/python:3.5
+      - image: circleci/python:3.6
         environment:
           TOXENV: lint
-  py35-core:
-    <<: *common
-    docker:
-      - image: circleci/python:3.5
-        environment:
-          TOXENV: py35-core
-  py36-core:
+  py36:
     <<: *common
     docker:
       - image: circleci/python:3.6
         environment:
-          TOXENV: py36-core
-  pypy3-core:
+          TOXENV: py36
+  py37:
     <<: *common
     docker:
-      - image: pypy
+      - image: circleci/python:3.7
         environment:
-          TOXENV: pypy3-core
+          TOXENV: py37
 workflows:
   version: 2
   test:
     jobs:
       - doctest
       - lint
-      - py35-core
-      - py36-core
-      - pypy3-core
+      - py36
+      - py37
