+++ docs/api.rst
@@ -7,9 +7,6 @@ Cancel Token
 .. autoclass:: cancel_token.CancelToken
   :members:
 
-.. autofuncion:: cancel_token.wait_with_token
-  :members:
-
 Exceptions
 ----------
