+++ docs/index.rst
@@ -3,6 +3,8 @@ asyncio-cancel-token
 
 Task cancellation pattern for ``asyncio`` applications.
 
+Inspired by https://vorpus.org/blog/timeouts-and-cancellation-for-humans/
+
 Contents
 --------
