+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist=
-    py{35,36,py3}-core
+    py{36,37,py3}
     lint
     doctest
 
@@ -22,12 +22,12 @@ ignore=
 [testenv]
 usedevelop=True
 commands=
-    core: pytest {posargs:tests/core}
+    py{36,py3}: pytest {posargs:tests}
     doctest: make -C {toxinidir}/docs doctest
 basepython =
     doctest: python
-    py35: python3.5
     py36: python3.6
+    py37: python3.7
     pypy3: pypy3
 extras=
     test
